requirements.txt
termcolor==2.4.0
selenium==4.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
WeasyPrint==62.3
attrs==24.2.0
lxml==5.3.0
lxml[html_clean]==0.2.2
twisted==24.7.0
pyee==11.1.1
pyppeteer==2.0.0
requests-html
cmake==3.29.6
dlib==19.24.6
numpy==1.26.3
opencv-python==4.9.0.80